rm internal
my idea was that my interface has roughly two parts:

 1. Iterator[T] and Make(), which implement a generic iterator
 2. a bunch of other functions that create iterators from common
    entities (slices, different map aspects, ...) and combine and
    consume iterators based on the primitives of 1.

so the idea of the internal package was to hide the really internal bits
of 1 from the bits of 2, partially so that I could feel I was honestly
consuming my basic iterator implementation and vaguely convince myself
it was a decent abstraction.

but in reality, I was really only hiding makeStop, which is not very
important (calling it as a Make() consumer doesn't exactly break
anything). so just don't.